+++ src/main/java/vn/edu/fpt/petworldplatform/repository/CustomerRepo.java
@@ -7,7 +7,7 @@ import vn.edu.fpt.petworldplatform.entity.Customer;
 import java.util.Optional;
 
 @Repository
-public interface CustomerRepo extends JpaRepository<Customer, Long> {
+public interface CustomerRepo extends JpaRepository<Customer, Integer> {
     Optional<Customer> findByUsername(String username);
 
     boolean existsByEmail(String email);
