+++ src/main/java/vn/edu/fpt/petworldplatform/repository/StaffRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface StaffRepository extends JpaRepository<Staff, Long> {
+public interface StaffRepository extends JpaRepository<Staff, Integer> {
 
     Optional<Staff> findByUsername(String username);
