+++ src/main/java/vn/edu/fpt/petworldplatform/service/BookingService.java
@@ -32,7 +32,7 @@ public class BookingService {
     private final PetRepo petRepo;
     private final ServiceItemRepository serviceItemRepository;
 
-    public List<Pets> findPetsByCustomerId(Long customerId) {
+    public List<Pets> findPetsByCustomerId(Integer customerId) {
         return petRepo.findByOwner_CustomerId(customerId);
     }
 
@@ -97,7 +97,7 @@ public class BookingService {
     }
 
     @Transactional
-    public Appointment createAppointment(Long customerId, Long petId, LocalDateTime appointmentDate,
+    public Appointment createAppointment(Integer customerId, Long petId, LocalDateTime appointmentDate,
                                          String note, List<Integer> serviceIds) {
         String code = generateAppointmentCode();
         
@@ -153,7 +153,7 @@ public class BookingService {
         return "APT-" + shortUuid.toUpperCase();
     }
 
-    public List<Appointment> findAppointmentsByCustomerId(Long customerId) {
+    public List<Appointment> findAppointmentsByCustomerId(Integer customerId) {
         return appointmentRepository.findByCustomerIdOrderByAppointmentDateDesc(customerId);
     }
 
@@ -162,7 +162,7 @@ public class BookingService {
         return appointmentRepository.findByCustomerIdAndStatusInOrderByAppointmentDateDesc(customerId, activeStatuses);
     }
 
-    public Optional<Appointment> findAppointmentByIdAndCustomerId(Integer id, Long customerId) {
+    public Optional<Appointment> findAppointmentByIdAndCustomerId(Integer id, Integer customerId) {
         return appointmentRepository.findById(id)
                 .filter(a -> a.getCustomerId().equals(customerId));
     }
@@ -176,7 +176,7 @@ public class BookingService {
         return appointmentServiceLineRepository.findAllByAppointmentIdsWithService(appointmentIds);
     }
 
-    public Optional<String> canCancelOrReschedule(Integer appointmentId, Long customerId) {
+    public Optional<String> canCancelOrReschedule(Integer appointmentId, Integer customerId) {
         Optional<Appointment> opt = findAppointmentByIdAndCustomerId(appointmentId, customerId);
         if (opt.isEmpty()) return Optional.of("Appointment not found.");
         Appointment a = opt.get();
@@ -191,7 +191,7 @@ public class BookingService {
     }
 
     @Transactional
-    public void cancelAppointment(Integer appointmentId, Long customerId, String reason) {
+    public void cancelAppointment(Integer appointmentId, Integer customerId, String reason) {
         Optional<String> err = canCancelOrReschedule(appointmentId, customerId);
         if (err.isPresent()) throw new IllegalArgumentException(err.get());
         Appointment a = findAppointmentByIdAndCustomerId(appointmentId, customerId).orElseThrow(() -> new IllegalArgumentException("Appointment not found"));
@@ -209,7 +209,7 @@ public class BookingService {
     }
 
     @Transactional
-    public void rescheduleAppointment(Integer appointmentId, Long customerId, LocalDateTime newStart) {
+    public void rescheduleAppointment(Integer appointmentId, Integer customerId, LocalDateTime newStart) {
         Optional<String> err = canCancelOrReschedule(appointmentId, customerId);
         if (err.isPresent()) {
             // Ensure we use the exact BR-18 message from user requirement
@@ -260,7 +260,7 @@ public class BookingService {
      * Delete a canceled appointment owned by customer.
      */
     @Transactional
-    public void deleteAppointmentIfCanceled(Integer appointmentId, Long customerId) {
+    public void deleteAppointmentIfCanceled(Integer appointmentId, Integer customerId) {
         Appointment a = appointmentRepository.findById(appointmentId).orElseThrow(() -> new IllegalArgumentException("Appointment not found"));
         if (!a.getCustomerId().equals(customerId)) throw new IllegalArgumentException("Unauthorized to delete this appointment.");
         if (!"canceled".equalsIgnoreCase(a.getStatus())) throw new IllegalArgumentException("Only canceled appointments can be deleted.");
